backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
joblib==1.4.2
scikit-learn==1.5.1
xgboost==2.1.1
shap==0.46.0
pandas==2.2.2
numpy==1.26.4
python-multipart==0.0.9
sqlalchemy==2.0.32
pymysql==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pymysql
cryptography
